Adds user authentication and management API
Implements user registration, login, and profile retrieval endpoints with JWT-based authentication.
Defines request/response schemas and validation rules using OpenAPI specification.
Includes exception handling and custom error responses.
Integrates password encryption for secure storage.
Adds tests for login and registration use cases and config classes.

+++ infrastructure/entry-points/reactive-web/src/test/java/co/com/pragma/api/UserApiHandlerTest.java
@@ -25,6 +25,7 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import reactor.core.publisher.Mono;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -83,7 +84,7 @@ class UserApiHandlerTest {
         validRequest.setEmail("jane.doe@example.com");
         validRequest.setIdNumber("98765432");
         validRequest.setPhone(3219876543L);
-        validRequest.setBaseSalary(3000000.0);
+        validRequest.setBaseSalary(BigDecimal.valueOf(3000000.0));
         validRequest.setBirthDate(LocalDate.of(2000, 1, 1));
         validRequest.setAddress("Avenida Siempre Viva 742");
         validRequest.setRole("ROLE_CLIENTE");
